Added check for OS when killing existing proces

+++ StartService.R
@@ -1,11 +1,16 @@
 
+# Determine OS
+OS <- Sys.info()["sysname"]
+
 #### Load libraries ####
 library(Rserve)
 
 #### Initiate server ####
 
-# Kill any existing instance of Rserve
-system("killall -INT Rserve")
+# If on Linux, Kill any existing instance of Rserve
+if (OS == "Linux"){
+  system("killall -INT Rserve")
+}
 
 # Start Rserve
 Rserve(args = "--no-save")
